Let error-analysis handle cronjob args, also Cronjob args can take more than one file pattern Remove unnecessary assertion.
