+++ log_analyzer/log_analyzer.py
@@ -1,21 +1,19 @@
 import argparse
-import datetime
 import importlib
 import ipaddress
 import operator
-import re
-from argparse import Namespace
-from pathlib import Path
 
-import pytz
 from typing import List, Optional, cast
 
 from abstract_configuration import AbstractConfiguration
+from cronjob_utils import convert_cronjob_to_batchjob
 from log_entry import LogReader
 from log_parser import LogParser
 from ip_to_host_converter import IpToHostConverter
 
+
 DEFAULT_FIELDS_PREFIX = 'https://tools.pds-rings.seti.org'
+
 
 def main(arguments: Optional[List[str]] = None) -> None:
     def parse_ignored_ips(x: str) -> List[ipaddress.IPv4Network]:
@@ -101,56 +99,6 @@ def main(arguments: Optional[List[str]] = None) -> None:
         elif args.fake_realtime:
             log_entries_list.sort(key=operator.attrgetter('time'))
             log_parser.run_realtime(iter(log_entries_list))
-
-
-def convert_cronjob_to_batchjob(args: Namespace) -> None:
-    if len(args.log_files) != 1:
-        raise Exception("Must specify exactly one log file pattern for cronjob mode")
-    log_file_pattern = args.log_files[0]
-    if '%' not in log_file_pattern:
-        raise Exception("Must specify a log file pattern, rather than a log file")
-
-    output_file_pattern = args.output
-    if not output_file_pattern:
-        raise Exception("Must specify the output file pattern for cronjob mode")
-    run_date = parse_cronjob_date_arg(args)
-    log_files = [datetime.datetime(year=run_date.year, month=run_date.month, day=day).strftime(log_file_pattern)
-                for day in range(1, run_date.day + 1)]
-    # Rob wants me to silently ignore non-existent files.
-    log_files = [file for file in log_files if Path(file).exists()]
-    output_file = run_date.strftime(output_file_pattern)
-    if log_files:
-        # Create all necessary intermediate directories
-        Path(output_file).parent.mkdir(parents=True, exist_ok=True)
-    args.log_files = log_files
-    args.output = output_file
-    args.batch = True
-
-
-def parse_cronjob_date_arg(args: Namespace) -> datetime.datetime:
-    '''Figure out the date to use, based on the --cronjob_date argument.'''
-    cronjob_date = args.cronjob_date
-    # if the argument isn't present, use today
-    if not cronjob_date:
-        today = datetime.datetime.now(tz=pytz.utc).replace(hour=0, minute=0, second=0, microsecond=0)
-        return today
-    # if the argument is -<number>, then it means that many days ago
-    match = re.match(r'-(\d+)', cronjob_date)
-    if match:
-        today = datetime.datetime.now(tz=pytz.utc).replace(hour=0, minute=0, second=0, microsecond=0)
-        return today - datetime.timedelta(days=int(match.group(1)))
-    # if the argument is dddd-dd, then it is a year and month, and indicates the last day of that month
-    match = re.match(r'(\d\d\d\d)-(\d\d)', cronjob_date)
-    if match:
-        year_month = datetime.datetime(tzinfo=pytz.utc, year=int(match.group(1)), month=int(match.group(2)), day=1)
-        sometime_following_month = year_month + datetime.timedelta(days=31)
-        return sometime_following_month - datetime.timedelta(days=sometime_following_month.day)
-    # if the argument is dddd-dd-dd, then treat it as year-month-day
-    match = re.match(r'(\d\d\d\d)-(\d\d)-(\d\d)', cronjob_date)
-    if match:
-        return datetime.datetime(
-            tzinfo=pytz.utc, year=int(match.group(1)), month=int(match.group(2)), day=int(match.group(3)))
-    raise Exception('cronjob_date must be one of -<int>, yyyy-mm, or yyyy-mm-dd')
 
 
 if __name__ == '__main__':
